+++ simple-stack/src/main/java/com/zhuinden/simplestack/navigator/Navigator.java
@@ -293,17 +293,6 @@ public class Navigator {
     }
 
     /**
-     * Returns if a given alias exists in a given scope.
-     *
-     * @param scopeTag the scope tag
-     * @param alias    the alias
-     * @return whether the alias exists
-     */
-    public static boolean hasAlias(@NonNull Context context, @NonNull String scopeTag, @NonNull String alias) {
-        return getBackstack(context).hasAlias(scopeTag, alias);
-    }
-
-    /**
      * Returns if a given scope exists.
      *
      * @param scopeTag   the scope tag